requirements.txt
# LangChain 教程依赖配置
# 更新日期：2026-03-16
# Python 版本要求：3.10+

# ==================== 核心依赖 ====================
langchain>=0.1.0
langchain-core>=0.1.7
langchain-community>=0.0.10

# ==================== 模型集成 ====================
# OpenAI
langchain-openai>=0.0.5
openai>=1.12.0

# Ollama (本地模型 - 推荐)
langchain-ollama>=0.1.0
ollama>=0.1.6

# 阿里云 DashScope
dashscope>=1.14.0

# ==================== 向量数据库 ====================
faiss-cpu>=1.7.4
chromadb>=0.4.22

# ==================== 文档处理 ====================
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# ==================== 数据处理 ====================
orjson>=3.9.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
tiktoken>=0.5.2
python-dotenv>=1.0.1

# ==================== 开发工具 ====================
pytest>=8.0.0

# ==================== Web 框架 (可选) ====================
fastapi>=0.109.0
uvicorn>=0.27.0
